Remove direct use of PgpSignatory in the Sign task
Conceptually, the signing plugin allows for using multiple signatories
that implement different signing methods. However, in practice the Sign
task always casts the signatory instance returned by Sign.getSignatory()
into a PgpSignatory. This makes it impossible to use a signatory that is
not derived from PgpSignatory.

The Sign task needs to access the PgpSignatory in order to retrieve the
key id which serves as task input value for the signatory property.

This commit fixes this direct dependency by adding the method
getInputValue() to the Signatory interface. This method returns a value
representing the input value of the signatory property. The Sign task
can then use this method for retrieving a value that represents the
signatory property for a specific signatory implementation.